fix: extract tests from src into tests so that they run as part of wasm-pack test
